+++ src/main/java/com/squad03/flap/repository/ListaRepository.java
@@ -3,5 +3,5 @@ package com.squad03.flap.repository;
 import com.squad03.flap.model.Lista;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ListaRepository extends JpaRepository<Lista, Integer> {
+public interface ListaRepository extends JpaRepository<Lista, Long> {
 }
